Updated postgresql permissions grants
