+++ cli/macrostrat/cli/database/__init__.py
@@ -4,6 +4,12 @@ from sys import exit, stderr, stdin, stdout
 from typing import Any, Callable, Iterable
 
 import typer
+from macrostrat.database import Database
+from macrostrat.database.transfer import pg_dump_to_file, pg_restore_from_file
+from macrostrat.database.transfer.utils import raw_database_url
+from macrostrat.database.utils import get_sql_files
+from macrostrat.utils import get_logger
+from macrostrat.utils.shell import run
 from pydantic import BaseModel
 from rich import print
 from sqlalchemy import make_url, text
@@ -11,19 +17,11 @@ from typer import Argument, Option
 
 from macrostrat.core import MacrostratSubsystem, app
 from macrostrat.core.migrations import run_migrations
-from macrostrat.database import Database
-from macrostrat.database.transfer import pg_dump_to_file, pg_restore_from_file
-from macrostrat.database.transfer.utils import raw_database_url
-from macrostrat.database.utils import get_sql_files
-from macrostrat.utils import get_logger
-from macrostrat.utils.shell import run
-
 from ._legacy import get_db
-
 # First, register all migrations
 # NOTE: right now, this is quite implicit.
 from .migrations import load_migrations
-from .utils import engine_for_db_name
+from .utils import engine_for_db_name, setup_postgrest_access
 
 log = get_logger(__name__)
 
@@ -396,6 +394,19 @@ def run_scripts(migration: str = Argument(None)):
 
 
 db_app.command(name="migrations", rich_help_panel="Schema management")(run_migrations)
+
+
+def update_permissions():
+    """Setup permissions for the PostgREST API.
+
+    NOTE: This is a stopgap until we have a better permssions system.
+    """
+    db = get_db()
+    setup_postgrest_access("macrostrat_api")(db)
+    db.run_sql("NOTIFY pgrst, 'reload schema';")
+
+
+db_app.command(name="permissions", rich_help_panel="Helpers")(update_permissions)
 
 
 ### Helpers
