+++ cli/macrostrat/cli/database/utils.py
@@ -2,15 +2,14 @@ from contextlib import contextmanager
 from typing import Optional
 from uuid import uuid4
 
+from macrostrat.database import Database
+from macrostrat.database.utils import run_query, run_sql
 from psycopg2.sql import Identifier
 from rich import print
 from sqlalchemy.engine import Engine, create_engine
 from sqlalchemy.engine.url import URL, make_url
 
 from macrostrat.core.config import settings
-from macrostrat.database import Database
-from macrostrat.database.utils import run_query, run_sql
-
 from ._legacy import get_db
 
 
@@ -336,3 +335,13 @@ def grant_schema_usage(
         """,
             params,
         )
+
+
+def setup_postgrest_access(schema: str):
+    """Run basic grant statements to allow PostgREST to access the schema"""
+
+    def run_updates(db):
+        grant_schema_usage(db, schema, "web_anon")
+        grant_schema_usage(db, schema, "web_user", tables=False, sequences=True)
+
+    return run_updates
